+++ packages/redux-oidc/src/authAdapter.ts
@@ -1,3 +1,5 @@
+import convertToCamelCase from 'camelcase-keys';
+
 import { removeTokens } from './utils';
 
 export interface AuthAdapterConfiguration {
@@ -33,37 +35,65 @@ export interface AuthenticationProviderUrls {
   revocationEndpoint: string;
   tokenEndpoint: string;
   userInfoEndpoint: string;
+  [key: string]: string | Record<string, string>;
 }
 
-export class AuthAdapater {
+/*  This class serves as an adapter for handling authentication.
+	This class uses Singleton patern for initialization and should be initilized with
+	getInstance method/
+*/
+export class AuthAdapter {
   private _configuration: AuthAdapterConfiguration;
+  private _oidcConfiguration?: AuthenticationProviderUrls;
+  private static _instance: AuthAdapter;
 
-  constructor(configuration: AuthAdapterConfiguration) {
+  private constructor(configuration: AuthAdapterConfiguration) {
     this._configuration = configuration;
   }
 
-  /**
-   * Using the configuration authority creates
-   * OIDC endpoints for different usecases
-   */
-  public getConfigurationEndpoints(): AuthenticationProviderUrls {
-    return {
-      authorizationEndpoint: `${this._configuration.authority}/protocol/openid-connect/auth`,
-      endSessionEndpoint: `${this._configuration.authority}/protocol/openid-connect/logout`,
-      revocationEndpoint: `${this._configuration.authority}/protocol/openid-connect/revoke`,
-      tokenEndpoint: `${this._configuration.authority}/protocol/openid-connect/token`,
-      userInfoEndpoint: `${this._configuration.authority}/protocol/openid-connect/userinfo`,
-    };
+  public static getInstance(configuration: AuthAdapterConfiguration): AuthAdapter {
+    if (!AuthAdapter._instance) {
+      AuthAdapter._instance = new AuthAdapter(configuration);
+    }
+    return AuthAdapter._instance;
   }
+
+  public async fetchOidcConfig() {
+    try {
+      const response = await fetch(new URL(this._configuration.authority + '/.well-known/openid-configuration'));
+
+      if (!response.ok) {
+        throw `OIDC config not available. Status: ${response.status} ${response.statusText}`;
+      }
+
+      if (response.body === null) {
+        throw `OIDC config empty. Response: ${response} (${response.status} ${response.statusText})`;
+      }
+
+      return convertToCamelCase(await response.json(), { deep: true });
+    } catch (error) {
+      if (error instanceof SyntaxError) {
+        // Unexpected token < in JSON
+        console.debug('Syntax error during OIDC config fetch', error);
+      }
+      throw new Error(String(error));
+    }
+  }
+
+  public getConfigurationEndpoints = async () => {
+    if (!this._oidcConfiguration) {
+      this._oidcConfiguration = await this.fetchOidcConfig();
+    }
+    return this._oidcConfiguration as AuthenticationProviderUrls;
+  };
 
   public getBaseUrl() {
     return this._configuration.baseUrl;
   }
 
-  public getLoginUrl() {
-    return `${this.getConfigurationEndpoints().authorizationEndpoint}?response_type=code&client_id=${
-      this._configuration.clientId
-    }&redirect_uri=${this._configuration.redirectUri}&scope=${this._configuration.scope}`;
+  public async getLoginUrl() {
+    const { authorizationEndpoint } = await this.getConfigurationEndpoints();
+    return `${authorizationEndpoint}?response_type=code&client_id=${this._configuration.clientId}&redirect_uri=${this._configuration.redirectUri}&scope=${this._configuration.scope}`;
   }
 
   public getSavedLocation = () => sessionStorage.getItem('saved_location') || undefined;
@@ -72,9 +102,9 @@ export class AuthAdapater {
   /**
    * Redirect user to sign in provider
    */
-  public startOidcSignIn(redirectUrl = window.location.pathname) {
+  public async startOidcSignIn(redirectUrl = window.location.pathname) {
     this.saveLocationForRedirect(redirectUrl);
-    window.location.replace(this.getLoginUrl());
+    window.location.replace(await this.getLoginUrl());
   }
 
   public getAccessToken = () => localStorage.getItem('access_token');
@@ -82,7 +112,8 @@ export class AuthAdapater {
   public getIdToken = () => localStorage.getItem('id_token');
 
   public async signOut(signOutUrl?: string): Promise<void> {
-    const logoutUrl = new URL(this.getConfigurationEndpoints().endSessionEndpoint);
+    const { endSessionEndpoint } = await this.getConfigurationEndpoints();
+    const logoutUrl = new URL(endSessionEndpoint);
     logoutUrl.searchParams.append('client_id', this._configuration.clientId);
 
     const idToken = localStorage.getItem('id_token');